move fight functionality into its own class
