+++ algobattle/battle_wrapper.py
@@ -10,10 +10,10 @@ import logging
 from abc import ABC, abstractmethod
 from typing import Any, Generator, Generic, Type, TypeVar
 from inspect import isabstract, signature, getdoc
+from algobattle.fight import Fight
 
 from algobattle.problem import Problem
 from algobattle.team import Team, BattleMatchups, Matchup
-from algobattle.docker import DockerError, DockerConfig
 from algobattle.util import parse_doc_for_param
 
 logger = logging.getLogger("algobattle.battle_wrapper")
@@ -35,9 +35,7 @@ class BattleWrapper(ABC, Generic[Instance, Solution]):
         if not isabstract(cls):
             BattleWrapper._battle_wrappers[cls.__name__.lower()] = cls
 
-    def __init__(
-        self, problem: Problem[Instance, Solution], docker_config: DockerConfig = DockerConfig(), **options: dict[str, Any]
-    ):
+    def __init__(self, problem: Problem[Instance, Solution], fight: Fight, **options: dict[str, Any]):
         """Builds a battle wrapper object with the given option values.
 
         Logs warnings if there were options provided that this wrapper doesn't use.
@@ -54,7 +52,7 @@ class BattleWrapper(ABC, Generic[Instance, Solution]):
             Dict containing option values.
         """
         self.problem = problem
-        self.docker_config = docker_config
+        self.fight = fight
 
     @classmethod
     def get_arg_spec(cls) -> dict[str, dict[str, Any]]:
@@ -95,10 +93,6 @@ class BattleWrapper(ABC, Generic[Instance, Solution]):
                 out[param.name] = kwargs
         return out
 
-    @classmethod
-    def check_compatibility(cls, problem: Problem, options: dict[str, Any]) -> bool:
-        return True
-
     @abstractmethod
     def wrapper(self, matchup: Matchup) -> Generator[Result, None, None]:
         """The main base method for a wrapper.
@@ -115,155 +109,6 @@ class BattleWrapper(ABC, Generic[Instance, Solution]):
             The Match object on which the battle wrapper is to be executed on.
         """
         raise NotImplementedError
-
-    def _one_fight(self, matchup: Matchup, instance_size: int) -> float:
-        """Execute a single fight of a battle between a given generator and solver for a given instance size.
-
-        Parameters
-        ----------
-        instance_size : int
-            The instance size, expected to be a positive int.
-
-        Returns
-        -------
-        float
-            Returns the approximation ratio of the solver against
-            the generator (1 if optimal, 0 if failed, >=1 if the
-            generator solution is optimal).
-        """
-        try:
-            instance, generator_solution = self._run_generator(matchup.generator, instance_size)
-        except ValueError:
-            return self.problem.approx_cap
-
-        try:
-            solver_solution = self._run_solver(matchup.solver, instance_size, instance)
-        except ValueError:
-            return 0
-
-        approximation_ratio = self.problem.approximation_ratio(instance, instance_size, generator_solution, solver_solution)
-        logger.info(
-            f"Solver of group {matchup.solver} yields a valid solution with an approx. ratio of {approximation_ratio}."
-        )
-        return approximation_ratio
-
-    def _run_generator(self, team: Team, instance_size: int) -> tuple[Instance, Solution]:
-        """Execute the generator of match.generating_team and check the validity of the generated output.
-
-        If the validity checks pass, return the instance and the certificate solution.
-
-        Parameters
-        ----------
-        instance_size : int
-            The instance size, expected to be a positive int.
-
-        Returns
-        -------
-        Instance, Solution
-            The generated instance and solution.
-
-        Raises
-        ------
-        ValueError
-            If no valid instance and solution can be generated.
-        """
-        if self.docker_config.space_generator is not None:
-            scaled_memory = self.problem.generator_memory_scaler(self.docker_config.space_generator, instance_size)
-        else:
-            scaled_memory = None
-
-        logger.debug(f"Running generator of group {team}...")
-        try:
-            output = team.generator.run(
-                str(instance_size),
-                timeout=self.docker_config.timeout_generator,
-                memory=scaled_memory,
-                cpus=self.docker_config.cpus,
-            )
-        except DockerError:
-            logger.warning(f"generator of team '{team}' didn't run successfully!")
-            raise ValueError
-
-        if not output:
-            logger.warning(f"No output was generated when running {team.generator}!")
-            raise ValueError
-
-        logger.debug("Checking generated instance and certificate...")
-        raw_instance, raw_solution = self.problem.split(output)
-        try:
-            instance = self.problem.parse_instance(raw_instance, instance_size)
-        except ValueError:
-            logger.warning(f"Generator {team} created a malformed instance!")
-            raise
-
-        try:
-            solution = self.problem.parse_solution(raw_solution, instance_size)
-        except ValueError:
-            logger.warning(f"Generator {team} created a malformed solution at instance size!")
-            raise
-
-        if not self.problem.verify_solution(instance, instance_size, solution):
-            logger.warning(f"Generator {team} failed due to a wrong certificate for its generated instance!")
-            raise ValueError
-
-        logger.info(f"Generated instance and certificate by group {team} are valid!")
-        return instance, solution
-
-    def _run_solver(self, team: Team, instance_size: int, instance: Instance) -> Solution:
-        """Execute the solver of match.solving_team and check the validity of the generated output.
-
-        If the validity checks pass, return the solver solution.
-
-        Parameters
-        ----------
-        team : Team
-            Solving team.
-        instance_size : int
-            The instance size, expected to be a positive int.
-        instance : Instance
-            The generated instance.
-
-        Returns
-        -------
-        Solution
-            The solver's solution.
-
-        Raises
-        ------
-        ValueError
-            If no solution can be generated
-        """
-        if self.docker_config.space_solver is not None:
-            scaled_memory = self.problem.solver_memory_scaler(self.docker_config.space_solver, instance_size)
-        else:
-            scaled_memory = None
-        instance_str = self.problem.encode_instance(instance)
-
-        logger.debug(f"Running solver of group {team}...")
-        try:
-            output = team.solver.run(
-                instance_str, timeout=self.docker_config.timeout_solver, memory=scaled_memory, cpus=self.docker_config.cpus
-            )
-        except DockerError:
-            logger.warning(f"Solver of team '{team}' didn't run successfully!")
-            raise ValueError
-
-        if not output:
-            logger.warning(f"No output was generated when running the solver of group {team}!")
-            raise ValueError
-
-        logger.debug("Checking validity of the solvers solution...")
-        try:
-            solution = self.problem.parse_solution(output.splitlines(), instance_size)
-        except ValueError:
-            logger.warning(f"Solver of group {team} created a malformed solution at instance size {instance_size}!")
-            raise
-
-        if not self.problem.verify_solution(instance, instance_size, solution):
-            logger.warning(f"Solver of group {team} yields a wrong solution at instance size {instance_size}!")
-            raise ValueError
-
-        return solution
 
     class Result:
         pass
